Add CMakeLists and main.cpp for Pivoting_QR project
A new CMakeLists.txt is added to set up the Pivoting_QR project. This includes necessary setup for linking to the OpenBLAS library and adding the main executable. A main.cpp file has also been created as the main entry point for the project. The code implements a QR decomposition method with pivoting.